(GH-14) Fix inconsistency in PPI::Token::HereDoc->heredoc().
Due to a bug in PPI::Token::Heredoc, $token->heredoc() does not include the
terminator line when the terminator is a simple bareword, but it includes the
terminator line in the other cases (single/double quoted , commands, escaped).

I wrote a patch to resolve this in PPI (see
https://github.com/adamkennedy/PPI/pull/71), but in the meantime this addresses
the inconsistency.