(GH-13) Require '&' to identify functions.
Ideally, I'll implement a full-fledged tokenizer to parse SQL safe comments.
But in the meantime, there's no way to distinguish between a space inside a
hash key and a space between a safe variable and a function, unless the
function comes with an ampersand (which are less likely to exist in hash keys).